+++ llm_thread/tests/__init__.py
@@ -3,4 +3,5 @@ from . import test_mail_message_store
 from . import test_llm_thread_tag
 from . import test_llm_thread_search
 from . import test_message_compute_author
+from . import test_maybe_generate_name
 from . import test_js